[CSI-1149]  - SDK - not notify error on bad ACL (system:error:permission)

+++ NexmoClient/NXMConversation.m
@@ -455,7 +455,7 @@
             self.uuid,
             self.name,
             self.displayName,
-            self.lastEventId,
+            (long)self.lastEventId,
             self.creationDate,
             self.myMember,
             self.allMembers];
